# server/src/main/resources/open-api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenFireStationManager API",
...
    "contact": {
      "name": "OpenFireStationManager",
      "url": "https://github.com/simonhauck/OpenFireStationManager"
    },
    "version": "v1"
  },
  "security": [ {
    "sessionCookie": [ ]
  } ],
...
  "paths": {
...
    "/api/admin/impressum": {
      "get": {
        "tags": [ "Admin - Legal" ],
        "summary": "Read the site notice for editing",
        "description": "Returns the current site notice (`Impressum`). Identical in content to the public endpoint; it exists separately so the admin UI can read it without depending on the public route.\n\nCall `impressumExists` first if you only need to know whether one is configured.\n\n**Authorisation:** requires the `ADMIN` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "getImpressum",
        "responses": {
          "200": {
            "description": "The current site notice.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ImpressumDto"
                }
              }
            }
          },
          "404": {
            "description": "No site notice has been configured yet.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [ "Admin - Legal" ],
        "summary": "Create or replace the site notice",
        "description": "Writes the site notice. At most one exists at a time, so this always **replaces** any previous notice rather than adding a second — there is no create/update distinction and no id to address.\n\nEvery field is applied as given, so send the complete notice, not just the parts you are changing.\n\n**Authorisation:** requires the `ADMIN` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "upsertImpressum",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpressumDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The stored site notice.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ImpressumDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. The `errors` array names each offending field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [ "Admin - Legal" ],
        "summary": "Remove the site notice",
        "description": "Deletes the site notice, after which the public endpoint reports it as absent. Succeeds even when no notice is configured, so it is safe to call blindly.\n\n**Authorisation:** requires the `ADMIN` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "deleteImpressum",
        "responses": {
          "204": {
            "description": "No notice is configured now."
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/public/impressum/exists": {
      "get": {
        "tags": [ "Legal" ],
        "summary": "Check whether a site notice is published",
        "description": "Returns a boolean without the notice itself, so a page can decide whether to show an Impressum link before fetching it.\n\n**Authentication:** none — this endpoint is public.",
        "operationId": "publicImpressumExists",
        "responses": {
          "200": {
            "description": "Whether a notice exists.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ImpressumExists"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [ ]
      }
    },
...
    "/api/admin/impressum/exists": {
      "get": {
        "tags": [ "Admin - Legal" ],
        "summary": "Check whether a site notice is configured",
        "description": "Returns a boolean without the notice itself. Use it to decide whether to call `getImpressum`, which fails with `404` when nothing is configured.\n\n**Authorisation:** requires the `ADMIN` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "impressumExists",
        "responses": {
          "200": {
            "description": "Whether a notice exists.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ImpressumExists"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": { }
          }
        }
      },
      "ValidationError": {
        "description": "A single field-level validation failure.",
        "properties": {
          "field": {
            "type": "string",
            "description": "Path of the offending field within the request body, e.g. `size`. Empty when the rule applies to the request as a whole rather than one field.",
            "example": "size"
          },
          "message": {
            "type": "string",
            "description": "Human-readable reason the value was rejected.",
            "example": "must not be blank"
          },
          "rejectedValue": {
            "type": "object",
            "description": "The value that was submitted and rejected. May be null."
          }
        }
      },
      "ValidationProblemDetail": {
        "description": "RFC 9457 problem document describing a request that failed validation, extended with an `errors` array naming each offending field.",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Invalid request"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Validation failed for request body"
          },
          "instance": {
            "type": "string",
            "description": "URI of the request that failed, when available."
          },
          "errors": {
            "type": "array",
            "description": "Every field-level failure detected in this request.",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "ImpressumDto": {
        "type": "object",
        "description": "The legally required site notice (`Impressum`). At most one exists at a time — writing replaces the previous one rather than adding to it.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the responsible organisation or person.",
            "example": "Freiwillige Feuerwehr Musterstadt",
            "maxLength": 255,
            "minLength": 0
          },
          "address": {
            "type": "string",
            "description": "Postal address, as a single free-text block.",
            "example": "Hauptstraße 1, 12345 Musterstadt",
            "maxLength": 1000,
            "minLength": 0
          },
          "contactEmail": {
            "type": "string",
            "format": "email",
            "description": "Public contact email address.",
            "example": "kontakt@feuerwehr-musterstadt.de",
            "maxLength": 255,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "description": "Public contact phone number. Optional.",
            "example": "+49 1234 567890",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [ "address", "contactEmail", "name" ]
      },
...
      "ImpressumExists": {
        "type": "object",
        "description": "Whether a site notice has been configured. Lets a caller check without needing permission to read the notice itself.",
        "properties": {
          "exists": {
            "type": "boolean",
            "description": "True when a site notice exists.",
            "example": true
          }
        },
        "required": [ "exists" ]
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Server-side session cookie issued by `POST /api/public/auth/login`. Send it on every authenticated request.",
        "name": "OFSM_AUTH",
        "in": "cookie"
      },
      "rememberMeCookie": {
        "type": "apiKey",
        "description": "Long-lived cookie issued when logging in with `rememberMe = true`, valid for 30 days. It authenticates on its own, so a long-lived client may send it instead of a session cookie.",
        "name": "OFSM_AUTH_REMEMBER_ME",
        "in": "cookie"
      }
    }
  }
}